+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/AbstractSpotlessMojo.java
@@ -47,6 +47,7 @@ import com.diffplug.spotless.Formatter;
 import com.diffplug.spotless.LineEnding;
 import com.diffplug.spotless.Provisioner;
 import com.diffplug.spotless.generic.LicenseHeaderStep;
+import com.diffplug.spotless.maven.antlr4.Antlr4;
 import com.diffplug.spotless.maven.cpp.Cpp;
 import com.diffplug.spotless.maven.generic.Format;
 import com.diffplug.spotless.maven.generic.LicenseHeader;
@@ -102,21 +103,14 @@ public abstract class AbstractSpotlessMojo extends AbstractMojo {
 	@Parameter
 	private Kotlin kotlin;
 
-	/** The XML extension is discontinued. */
-	@Parameter
-	@Deprecated
-	private com.diffplug.spotless.maven.xml.Xml xml;
-
 	@Parameter
 	private Cpp cpp;
 
 	@Parameter
 	private Typescript typescript;
 
-	/** The CSS extension is discontinued. */
 	@Parameter
-	@Deprecated
-	private com.diffplug.spotless.maven.css.Css css;
+	private Antlr4 antlr4;
 
 	@Parameter(property = "spotlessFiles")
 	private String filePatterns;
@@ -204,7 +198,7 @@ public abstract class AbstractSpotlessMojo extends AbstractMojo {
 	}
 
 	private List<FormatterFactory> getFormatterFactories() {
-		return Stream.concat(formats.stream(), Stream.of(java, scala, kotlin, cpp, typescript, css, xml))
+		return Stream.concat(formats.stream(), Stream.of(java, scala, kotlin, cpp, typescript, antlr4))
 				.filter(Objects::nonNull)
 				.collect(toList());
 	}